+++ db/models/contracts.js
@@ -4,11 +4,11 @@ const contracts = {
     readAll: async() => {
         return (await client.query('SELECT * FROM contracts'));
     },
-    create: async({ userId, companyId, contract, contractStatus }) => {
+    create: async({ userId, companyId, postId, contract, contractStatus }) => {
 
-        const SQL =`INSERT INTO contracts ("userId", "companyId", contract, "contractStatus") values ($1, $2, $3, $4) RETURNING *`;
+        const SQL =`INSERT INTO contracts ("userId", "companyId", "postId", contract, "contractStatus") values ($1, $2, $3, $4, $5) RETURNING *`;
 
-        return (await client.query(SQL, [ userId, companyId, contract, contractStatus ])).rows[0];
+        return (await client.query(SQL, [ userId, companyId, postId, contract, contractStatus ])).rows[0];
     }
 }
